requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
banks==2.1.2
beautifulsoup4==4.13.4
bitsandbytes==0.42.0
blake3==1.0.4
bleach==6.2.0
bm25s==0.2.14
Brotli==1.1.0
brotlipy==0.7.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
click==8.1.8
cloudpickle==3.1.1
codetiming==1.4.0
colorama==0.4.6
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.9.3
contourpy==1.3.2
cupy-cuda12x==13.4.1
cycler==0.12.1
dashscope==1.23.2
dataclasses-json==0.6.7
datasets==3.5.1
debugpy==1.8.14
decorator==5.2.1
deepspeed==0.16.9
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
faiss-gpu-cu12==1.11.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fastrlock==0.8.3
ffmpy==0.6.2
filelock==3.18.0
filetype==1.2.0
flash-attn==2.7.0.post2
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.0
gguf==0.16.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.1
griffe==1.7.3
groovy==0.1.2
grpcio==1.67.1
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.35.3
hydra-core==1.3.2
importlib_metadata==8.0.0
infinity_client==0.0.76
interegular==0.3.3
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.6
isoduration==20.11.0
isort==6.0.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
json5==0.12.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
langchain==1.1.2
langchain-core==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.4.52
lark==1.2.2
lightning-utilities==0.14.3
llama-cloud==0.1.19
llama-cloud-services==0.6.21
llama-index==0.12.34
llama-index-agent-openai==0.4.7
llama-index-cli==0.4.1
llama-index-core==0.12.34.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.21
llguidance==0.7.19
llvmlite==0.44.0
lm-format-enforcer==0.10.11
loralib==0.1.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.14.1
mdurl==0.1.2
milvus-lite==2.4.12
mistral_common==1.5.4
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
notebook==7.4.1
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.76.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.3
optimum==1.24.0
orjson==3.10.18
ormsgpack==1.12.1
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.3
pydantic-settings==2.10.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pymilvus==2.5.8
pynvml==12.0.0
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-agent==0.0.21
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.3
rpds-py==0.24.0
rsa==4.9.1
ruff==0.14.0
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
shellingham==1.5.4
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.46.2
striprtf==0.0.26
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensordict==0.6.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
torchdata==0.10.0
torchmetrics==1.7.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==3.2.0
typer==0.15.3
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
uv==0.9.22
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.30.0
vllm==0.8.5
wandb==0.19.10
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.25.0
